Fix infinite loop bug.
Flickr seems to have changed the behavior of the flickr.people.getPhotos API call slightly. When fetching a page beyond the user's highest numbered page, they previously returned an empty response with no photos. Now they seem to return the same response as the highest numbered page.

I changed the loop logic to stop looping when the "pages" attribute in the response is equal to the page that was just fetched.