Add support for downloading videos
Before this change, when a video was encountered, only a JPEG containing the
first frame of the video was downloaded.

With this change in place, full videos are downloaded in their original
resolution, into properly named files.

On failed video download, tell user via stderr to do it manually.

Some videos can be downloaded automatically, some can't, do our best.

We can't download videos that Flickr says are "being processed". But even in
that case, a user logged in in a browser can do it manually, so we just tell
them.