Set file timestamps on downloaded photos
This change also adds unit tests and instructions on how to run them (just do
`py.test`).

The rationale for this change is that I wanted to copy photos and videos from
Flickr into Google Photo. And for Google Photo to know when different photos
have been taken, I'm guessing that it needs correctly timestamped files.

To be able to get `py.test` to run the tests, the main script had to have a
`.py` file extension, so I added that.

This change has been tested on 33GB of photos and videos.