Rename exception and change superclass.
It seemed a little weird to inherit from IOError. The docs for IOError give me the impression that it should be used when a lower-level file operation fails. It's often used with an errno value. In our usage an HTTP request returned a non-200 response, which seems a little less "exception-y." Ineriting from Exception seems slightly simpler and more correct.

Also rename from VideoDownloadException to VideoDownloadError. I think this is the more common naming scheme.