Small change to force album dirs to be recreated.
The next time the script is run. The reason for this is that the directories might have invalid symlinks to the bug fixed in the previous commit. Alternatively I could have written code to look for and fix invalid symlinks, but that's a lot of work. Deleting/recreating the album dirs should be pretty harmless.

I should add that it has been difficult to fix this. I've been getting errors the last two days while trying to mirror my videos using the script. Hopefully temporary?