Add the clean option
This option is disabled by default. Only when enabled the files and folders that no longer exist in Flickr are deleted.
When both clean and ignore-photos are used, all the photos from the Flickr backup will be deleted (because information about photos is not taken from Flickr).
When both clean and ignore-videos are used, all the videos from the Flickr backup will be deleted (because information about videos is not taken from Flickr).