Improve logic.
I find this to be a little easier to reason about. It seems more direct to me. "If (this is a photo and we're not ignoring photos) or (this is a video and we're not ignore videos) then <do something>." The old logic was something like "If (this is a video and we're ignoring photos) or (this is a photo and we're ignoring videos) or (we're not ignoring anything) then <do something>."