Reinstate two commented out sys.exit(1) calls.
The script exits if there's a download error. I think it would be better if the script was more resilient to download errors and made a best-effort attempt to continue even if there's an error, but I'm not quite comfortable making that change yet. I added a TODO item to README.md.