Don't ask to download videos due to no metadata.
There's a bug right now where the script continues asking the user to download videos even after they've been downloaded. The problem is that the metadata file doesn't exist initially, which causes the script to think it needs to download the video again, which fails.

My easy solution is to use absense of the metadata file as an indication to download the file. A better solution would be to continue using absense of the metadata file as an indication to download the file (revert this commit), and instead change the script to save metadata when prompting the user to download the video. We'll have to change the metadata filename in this case, because we won't know the extension. I don't think it's too important, but this approach would make the script more likely to catch the scenario and re-download a file if the file was changed and the metadata file was deleted.